django-rest-api/requirements.txt
asgiref==3.8.1
Django==4.2.3
debugpy
sqlparse==0.5.0
psycopg2-binary==2.9.6
gunicorn==21.2.0
djangorestframework==3.14
django-filter==24.2